Add ability to find tests with different prefixes
Some testing frameworks for Swift / Objective-C use different test prefixes to identify tests. For instance, [Quick](https://github.com/Quick/Quick) uses the `spec` method name prefix instead of `test`. In order to be able to use the test-center fastlane plugin with `Quick`, we need to be able to identify these differently-named tests, and so this commit introduces the ability to override the prefix that XCTestList uses to find tests.